requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
et-xmlfile==1.0.1
gprof2dot==2019.11.30
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.0
openpyxl==3.0.6
pandas==1.2.1
psycopg2==2.8.6
pycodestyle==2.6.0
Pygments==2.7.4
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
whitenoise==5.2.0
xlrd==2.0.1

aiohttp~=3.7.4